+++ org.polymap.rap.openlayers/src/org/polymap/rap/openlayers/source/TileWMSSource.java
@@ -76,7 +76,7 @@ public class TileWMSSource
      * this. See http://mapserver.org/output/tile_mode.html.
      */
     @Concern(OlPropertyConcern.class)
-    public Config2<TileWMSSource,Double>            gutter;
+    public Config2<TileWMSSource,Float>            gutter;
 
     /**
      * Logo.
